feat: organize list-items in the same TS files as their containing lists
